Worked on authenticateUser(), updateStock(), fetchRentals(), added pushRental() to be called whenever a User rents an Item so it can be written to the rental DB.
Need to figure out how we want to return fetches, are we just creating a
new object instance of whatever we end up fetching?
For example when we fetch a rental from the DB, all we get are three
Strings. Do we need to instantiate a new Item object based on the
fetched itemID and create a new Rental object using the userID provided
to the method (assuming it comes from a user object), the Item object,
and the Date object?